requirements.txt
fastapi
uvicorn
qrcode
Pillow
pyzbar
python-multipart
requests
